Remove unused etcd code
