+++ go.mod
@@ -8,7 +8,6 @@ require (
 	github.com/Financial-Times/service-status-go v0.0.0-20160323111542-3f5199736a3d
 	github.com/Financial-Times/uuid-utils-go v0.0.0-20170516110427-e22658edd0f1
 	github.com/Sirupsen/logrus v0.11.2
-	github.com/coreos/etcd v3.1.2+incompatible
 	github.com/davecgh/go-spew v1.1.0 // indirect
 	github.com/giantswarm/retry-go v0.0.0-20151203102909-d78cea247d5e
 	github.com/gorilla/context v1.1.1 // indirect
@@ -21,9 +20,7 @@ require (
 	github.com/satori/go.uuid v1.1.0
 	github.com/stretchr/objx v0.0.0-20150928122152-1a9d0bb9f541 // indirect
 	github.com/stretchr/testify v1.1.4
-	github.com/ugorji/go v0.0.0-20170215201144-c88ee250d022 // indirect
 	github.com/willf/bitset v1.0.1-0.20161202170036-5c3c0fce4884 // indirect
-	golang.org/x/net v0.0.0-20170227185932-906cda9512f7
 	golang.org/x/sys v0.0.0-20170329061634-9a7256cb28ed // indirect
 	gopkg.in/check.v1 v1.0.0-20190902080502-41f04d3bba15 // indirect
 )
